test/dashboard: only test enabled python bindings
we should test the python version of MGR_PYTHON_VERSION_MAJOR, if we are
testing both py2.7 and py3, we are preparing for distributing dashboard
as a separated package independent of ceph-mgr.

restructure the dashboard tests as 2 matrices, so we have 2*2 tests for
coverage and lint respectively.

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 213652aa6df094d339f57b2b0d9616507a77678c)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>