test/dashboard: specify workdir using tox.ini
--workdir was introduced in tox v2.4.0, but the tox shipped by
ubuntu/xenial is 2.3.1. also because i removed the step to prepare the
tox using "pip install -r requirements.txt", what we have is the tox
v2.3.1 . so, instead of passing workdir in the command line, we specify
this setting in tox.ini.

also drop the tox in requirements.txt, it's just misleading. and we are
not using the tox installed in the venv.

Fixes: http://tracker.ceph.com/issues/23709
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit a165464930d2814daa9a04ade99d13648ee427ed)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>