mgr/dashboard: silence E741
This is a pretty questionable check because it complains
about the caller of an API instead of the API itself, if
one of the API's members/arguments is one of the
forbidden variable names such as 'O'.

The interface to pyopenssl includes an 'O' member
on the certificate object.

Signed-off-by: John Spray <john.spray@redhat.com>
(cherry picked from commit 1e768b3f4e5cce3493559a2ee6aa6a4aea6a4603)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>