mgr/dashboard: Add table actions component
With the help of this component it is easy to create all actions for a
table with the right permission for each table action with a simple
configuration.

Per default action button and action drop down list are shown. All
actions will be filtered depending on the permissions on
initialization. Per default all actions that have the permission
'update' or 'delete' are only enabled if an item is selected which
holds no running task.

You can use change the default behavior, to show actions only in a
named drop down.
For each item you can change it's default behaviors by using custom
functions to enable/disable an action or show/hide it
as action button or drop down item.

Please look at the tests to understand how to use it in practice.

Signed-off-by: Stephan Müller <smueller@suse.com>
(cherry picked from commit ef7a0de5cb139243d49a369b2810b268896964d6)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>