mgr/dashboard: Catch LookupError when checking the RGW status
This exception is thrown by the RGW client function _determine_rgw_addr(). Without catching that exception a 'The server encountered an unexpected condition which prevented it from fulfilling the request.' will be displayed in the UI instead of displaying a notification panel with a hint that the Object Gateway is not running.

Fixes: https://tracker.ceph.com/issues/35921

Signed-off-by: Volker Theile <vtheile@suse.com>
(cherry picked from commit 0035bb0831ee94293ec236fa8bb818a852ca0081)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>