Fix package{-lock}.json
We added a couple of new dashboard v2 dependencies but we did not update
package-lock.json for them. That makes the dashboard build fail since we
are using npm ci that checks the consistency of the package-lock.json
file instead of npm install that does not. I have generated this change
with npm install to avoid this issue and to allow ceph to build
properly.

Signed-off-by: Boris Ranto <branto@redhat.com>