mgr/dashboard: add backend-driven UI tables
As an example this PR displays the list of Ceph Auth users, keys and
caps.

It tries to minimize the amount of UI code required by feeding a generic
table-like component (RestTable) with backend-generated JSON data.

This is just a proof of concept and there's lot of room for improvement.

Fixes: https://tracker.ceph.com/issues/52701

Signed-off-by: Ernesto Puerta <epuertat@redhat.com>
Signed-off-by: Pere Diaz Bou <pdiazbou@redhat.com>
Signed-off-by: Nizamudeen A <nia@redhat.com>