rgw: Create base class for broken out SyncShard coroutines
The problem:

RGWDataSyncShardCR combines full and incremental sync in such a way
that yielding in full sync after the transition to incremental (but
before all the finishing up is done) can jump straight into
incremental.

To deal with this break full and incremental sync into their own
coroutine classes. For the shared functionality (but /not/ shared
data) between them, we have a base class into which parts of the
original class are moved.

Data that are shared between the two (shard cache, lease) are owned
and managed by RGWDataSyncShardCR, with the full and incremental base
class getting references.

Fixes: https://tracker.ceph.com/issues/57063
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>