pybind/mgr: tox.ini remove redundant `tox` env
Fixes: https://tracker.ceph.com/issues/57153

The envlist contained an environment named `lint`. There was no specific
customization of the lint testenv so it is essentially the same as
running the `py3` testenv.

This was probably a typo and was meant to be `pylint`. Unfortunately,
the pylint test env does not appear to work, probably because it was
never run as part of any automation. At the risk of leaving old stuff
behind I'm not removing the pylint testenv at the moment, only the
`lint` item in order to not run redundant tests.

Signed-off-by: John Mulligan <jmulligan@redhat.com>