osd/OSD: trim_maps - remove nrecieved
The intention of `nrecieved` was to guarantee that we will trim [first, last) of
the current MOSDMap being handled. However, this is not true anymore since we trim based on cluster_osdmap_trim_lower_bound.

For instance, we might begin trimming when we receieve a MOSDMap with first epoch of 799 and last of 800 and cluster_osdmap_trim_lower_bound of 200.
If our superblock's oldest_map is 1 (no trimming occurred yet), we are allowed to trim up to 200.
There is no relation between the current MOSDMap handled amount of epochs receieved
(last - first + 1) to the epoch where we allowed to trim to.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>