cephadm: split DaemonSubIdentity class off from DaemonIdentity
Split DaemonSubIdentity into it's own subclass of DaemonIdentity
in order to better use the type system for validation and clarity.
DaemonIdentity loses the `subcomponent` field so that there's no
mistaking a DaemonIdentity with a DaemonSubIdentity.
This will make a lot more sense once there are more uses of
DaemonSubIdentity in the codebase.

Signed-off-by: John Mulligan <jmulligan@redhat.com>