cephadm: split get_unit_name into two functions
The previous version of get_unit_name was effectively overloaded,
so that you could pass either an fsid, daemon_type, & daemon_id OR
pass an fsid and a "daemon_type" only that wasn't actually a daemon_type
but rather the instance name of the systemd service. Separate the
two use cases into two functions, making the naming more correct and
the code more audit-able overall.

Signed-off-by: John Mulligan <jmulligan@redhat.com>