test/librbd/fsx: fix wnbd check
When using WNBD, the IO size must be a multiple of the sector size
(WNBD only supports 512 at the moment).

We're currently checking the "wnbd_operations" variable, however
it's only defined on Windows, leading to a compilation failure
on other platforms.

We'll switch back to a separate boolean variable called "wnbd_disk".

Fixes: https://tracker.ceph.com/issues/62475
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>