cephadm: black format net_utils.py
This contains one small manual change reversing the black formatting.
According to the black docs [1], flake8's E203 is not pep8 compliant but
since flake8 is currently mandatory and black formatting is not we'll
leave it in a form that makes flake8 happy and attempt to resolve the
disagreement some time in the future.

[1]: https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices

Signed-off-by: John Mulligan <jmulligan@redhat.com>
Pair-programmed-with: Adam King <adking@redhat.com>
Co-authored-by: Adam King <adking@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>