osd: fix: slow scheduling when item_cost is large
We use the iops and bandwidth tested by
`ceph tell osd.0 bench 10737418240 204800 204800 100`
to verify the QoS function. iops was 400 and bandwidth was 80MiB/s.
When osd_mclock_scheduler_client_lim is set to 1,
the sequential write bandwidth is only half of the capacity.
Therefore, we believe that it should not unconditionally increase
osd_bandwidth_cost_per_io for each IO, but take the maximum of the two.

Fixes: https://tracker.ceph.com/issues/62812
co-author: yanghonggang <yanghonggang_yewu@cmss.chinamobile.com>
co-author: zhangjianwei <zhangjianwei2_yewu@cmss.chinamobile.com>
Signed-off-by: Jrchyang Yu <yuzhiqiang_yewu@cmss.chinamobile.com>