cephadm: add supports_split_cgroups property to Podman
Add a new method, supports_split_cgroups, to the Podman type. This
function returns a boolean indicating the the podman instance in
use supports split cgroups. This function will be later used to replace
various instances of `podman.version >= CGROUPS_SPLIT_PODMAN_VERSION`.
This encapsulates the check and allows it to be more easily modified
in the future.
It is also shorter, which is nicer to read and to type.

Signed-off-by: John Mulligan <jmulligan@redhat.com>