cephadm: move logic to build pull command to container engines
Move the logic needed to build a pull command to a function in the
container_engines.py file. This continues the effort to improve the
locality of items with regards to specific options for container
engines. I had thought to move the whole pull function but the logging
and retry code made me change my mind. We can always move it later if it
makes sense.

Signed-off-by: John Mulligan <jmulligan@redhat.com>