cephadm: always pass ctx to customize_container_{binds,mounts}
These functions often derive the binds and/or mounts from the context
variable. Thus we should have the base class method accept the context.
Not all subclassess will use it but it will be there for those that do.
Also, fix the type for customize_container_mounts - it should be a dict
not a list.

Signed-off-by: John Mulligan <jmulligan@redhat.com>