cephadm: mock a call to shutil.rmtree
Mock out calls to shutil.rmtree in test_deploy_and_rm_iscsi. The test
never asserts that things in the file system are removed, so making the
call a no-op is safe. Calling shutil.rmtree with the fake file system
enabled recently became a problem.  In python 3.12 (or 3.11?) changes to
the shutil module function shutil.rmtree are incompatible with the fake
file system that cephadm tests rely upon heavily.

Part of an effort to get ceph tox environments passing on Python 3.12.

Signed-off-by: John Mulligan <jmulligan@redhat.com>