mgr/cephadm: adding the SSL cert as a dependency for ingress service
This way when the user changes the SSL certificate in the spec and
performs an 'orch apply' cephadm will detect the change and
reconfigure the ingress service automatically. In addition we are also
adding the ssl_key because it was not included in the generated
haproxy configuration.

Signed-off-by: Redouane Kachach <rkachach@ibm.com>
(cherry picked from commit 8dbc5f84265035abe9f74aacbbeab2cc887359cc)

Conflicts:
	src/pybind/mgr/cephadm/services/ingress.py

Resolves: rhbz#2344352