rgw: set umask for luarocks temporary directory
mkdtemp creates the temporary directory with 0700 permissions by default, which restricts access to the user that created it. Since RGW can start as root and later drop privileges to the ceph user, the installed Lua code needs to be accessible by other users.

This commit explicitly sets the umask (chmod) to ensure the temporary directory has appropriate permissions (755), allowing access to all users.

Signed-off-by: Ansgar Jazdzewski <ansgar.jazdzewski@hetzner-cloud.de>