librbd: tolerate image not existing in ImageRemoveRequest
ImageRemoveRequest may be called when the image no longer exists (or
even never existed in case of a clone image whose creation is pending
on its parent showing up on the secondary) to clean up leftover mirror
metadata.  Upon failure to get the group spec on a non-existing image
header, the state machine should be advanced to remove_mirror_image().

This fixes sporadic failures in "TEST: cloned images" and "TEST: check
if removed images' OMAP are removed".

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>

Resolves: rhbz#2089305