mgr/cephadm: add options for haproxy to verify backend ssl cert
This requires us to have the CA cert used to sign that
certificate. In the case where we generate our own cert
for the ingress service, that's just our cephadm CA cert.
Otherwise, we'll need the user to pass in the CA cert
themselves for us to use for verification.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 8885dcba7ad77ab70d20b8b6632b4160de445843)

Conflicts:
	src/pybind/mgr/cephadm/services/ingress.py
	src/python-common/ceph/deployment/service_spec.py

Resolves: rhbz#2345489