pybind/mgr: add default set of cmake/ctest labels to tests
Add a pair of default test labels (`PurePython` and `PythonTox`) to the
tests generated by the add_tox_test cmake function. In addition, the
function can extend the set of labels applied with a new LABELS argument
to the cmake function.

Using the label can be done with something like:
```
ctest -L PythonTox -j8
```

Note that these tests are labeled PurePython because they do not
require C extensions to be built. The above example command can
be run successfully immediately after configuring, without having
to build the C++ sources first.

Signed-off-by: John Mulligan <jmulligan@redhat.com>