cephadm/tests: fix _dist_info function logic error
The _dist_info helper function had a logic error where it was checking
if 'entry.startswith(entry)' instead of 'entry.startswith(name)'. This
caused the function to always evaluate incorrectly when checking for
.dist-info or .egg-info entries in the zipapp.

This bug was preventing the test assertions from properly validating
that package metadata directories are included in the built cephadm
zipapp.

Fixes a bug introduced in commit 4be4a95df42.

Signed-off-by: Kefu Chai <k.chai@proxmox.com>