qa: Use the pythonic way to install CBT python requirements
The qa/tasks/cbt.py task currently uses a hard-coded list of OS python packages to install ny pre-requisite python packages for CBT. This means that if there is a change to the pre-requisite python packages in the CBT project it will not be picked up by teuthology unless the cbt.py task is also changed.
The pythonic way to install pre-requisite packages is to use a requirement.txt file for the project, and install these via the pip python package manager. This way when the CBT project is updated we do not also need to make changes to teuthology code.

Signed-off-by: Chris Harris <harriscr@uk.ibm.com>