ceph-volume: remove exc_info from ceph.conf load warning
This commit removes exc_info=1 from the logger.warning call when failing to
load ceph.conf. According to the preceding comment, this scenario can happen
legitimately, so it is not an unexpected error, it means there is no need to
clutter the logs with a full python traceback for this case.

Signed-off-by: Guillaume Abrioux <gabrioux@ibm.com>