qa/cephfs: give more time to tests in test_clone_stats.py
Currently tests wait for 10 seconds for clone progress bar to appear and
it usually takes 8 seconds to find it. In case of the host where test is
being run is slow, the progress bar can take more time to appear and the
test can fail unnecessarily. To avoid this, increase the waiting
duration.

Signed-off-by: Rishabh Dave <ridave@redhat.com>