cmake: build static seastar for release builds
When BUILD_SHARED_LIBS is set, seastar inherits this setting from the
parent CMake project, causing crimson to link against libseastar.so.
While this works in development environments, it breaks package
installation because libseastar.so is not included in the distribution:

```
  can't install ceph-crimson-osd:
    - nothing provides libseastar.so()(64bit) needed by
      ceph-crimson-osd-2:20.2.0-2.fc44.x86_64

  can't install ceph-osd:
    - nothing provides libseastar.so()(64bit) needed by
      ceph-osd-2:20.2.0-2.fc44.x86_64

  can't install ceph-test:
    - nothing provides libseastar.so()(64bit) needed by
      ceph-test-2:20.2.0-2.fc44.x86_64
```

Force seastar to build as a static library regardless of the parent
project's BUILD_SHARED_LIBS setting. This fixes the packaging issue
and provides a modest performance improvement by eliminating PLT/GOT
indirection overhead for seastar function calls.

Fixes: https://tracker.ceph.com/issues/74138

Signed-off-by: Kefu Chai <k.chai@proxmox.com>